Add files via upload
These files contain code for problem set 4 which was building the Kevin Bacon game in a way that user inputs could be put in and different simulations could be run in the game. 

The /bacon/ file directory contains the input files needed to build and run the game. The /PS4Final/ directory contains the code, an analysis of test cases, and a discussion of boundary inputs.

We have not modified the code since we wrote it for class because we believe that although we know better now, the code as is represents a point in our learning experience that we want to record--it works well, but could be made more efficient. Future programs will embody this efficiency.